Moves SoftwareStack component to ui folder
