+++ components/ui/SoftwareStack.vue
